Security update

+++ investment/src/main/java/my/investment/fd/Controllers/DeleteController.java
@@ -15,6 +15,7 @@ import org.springframework.web.server.ResponseStatusException;
 import my.investment.fd.Classes.Role;
 import my.investment.fd.Entities.FixedDeposit;
 import my.investment.fd.Repositories.FixedDepositRepository;
+import my.investment.fd.Security.AuthUtil;
 
 
 
@@ -39,6 +40,10 @@ public class DeleteController {
     ) {
         FixedDeposit fd = fixedDepositRepository.findById(id).orElse(null);
         if (fd == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Non-existent Fixed Deposit ID: " + id);
+
+        if ( !AuthUtil.hasCRUDPermission(fd) ) 
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Only admin and owner can delete this Fixed Deposit");
+
         fixedDepositRepository.delete(fd);
         return ResponseEntity.ok(fd);
     }
